% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscan}
\alias{dbscan}
\title{This is an implementation of the DBSCAN Algorithm}
\usage{
dbscan(data, eps, minPts)
}
\arguments{
\item{data}{The data to be clustered.}

\item{eps}{The distance defining the neighborhood of a point.}

\item{minPts}{The minimum point count needed to form a dense region.}
}
\description{
We're using bfs to find the connected parts
}
