% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{extract_dbscan}
\alias{extract_dbscan}
\title{extract_dbscan}
\usage{
extract_dbscan(optics_result, eps_prime = optics_result$eps)
}
\arguments{
\item{optics_result}{Uses the result of the OPTIC algorithm (see 'optics')}

\item{eps_prime}{(optional) value to extract clusters, default is \code{eps} of optics_result.}
}
\value{
Returns a list of 2 elements \code{ordered_labels} and \code{labels}
}
\description{
Extract clusters from the result of OPTICS algorithm. Clusters are found by 'horizontally cutting' the
reachability plot with an eps_prime value.
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,1,0,1,2,0),ncol=2), 0.3)
eps <- 0.5
minPts <- 3
#Obtain an optics result
optics_result <- optics(data, eps, minPts)
#execute function with default eps_prime
extract_dbscan(optics_result)

}
