% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{gausskernel}
\alias{gausskernel}
\title{Algorithmus: spektrales Clustering}
\usage{
gausskernel(x, y)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthält einen Messwert in R^d}

\item{num_cluster}{int. Anzahl der Cluster}

\item{f_kernel}{(optional) function.Kernfunktion. Default: Gauss-Kern mit Abstieg h}

\item{dim}{(optional) int. Dimension der erstellten k-dimensionalen Projektionen durch spektrales Clustern. Default: 2}
}
\value{
k-dimensionale Projektionen (selbes Format wie data)
}
\description{
berechnet k-dimensionale Projektionen zu gegebenen Daten
}
