% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_clusters.R
\name{gen_clusters}
\alias{gen_clusters}
\title{gen_clusters}
\usage{
gen_clusters(n, means, deviation)
}
\arguments{
\item{n}{Anzahl an Punkten pro Cluster}

\item{means}{Mittelpunkte der Cluster}

\item{deviation}{Standardabweichung bei Lärm}
}
\value{
data
}
\description{
gen_clusters
}
