% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchal-clustering.R
\name{hierarchial_clustering}
\alias{hierarchial_clustering}
\title{This is an implementation of a Hierarchical Clustering Algorithm. This
implementation is agglomerative / bottom-up}
\usage{
hierarchial_clustering(data, metric = "euclidian", type = "complete")
}
\description{
This is an implementation of a Hierarchical Clustering Algorithm. This
implementation is agglomerative / bottom-up
}
