% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{k_means_predict}
\alias{k_means_predict}
\title{k_means_predict}
\usage{
k_means_predict(x, means)
}
\arguments{
\item{x}{atomic vector or matrix with measurements}

\item{means}{cluster means (e.g. from k-means algorithm)}
}
\value{
predicted labels of new measurements
}
\description{
k_means_predict
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,0,1,2), ncol=2), 0.3)
 clustering <- k_means(data,3)
 k_means_predict(c(1.2,0.8), clustering$means)

data <- gen_clusters(100,matrix(1:30, ncol=6),0.2)
result <- k_means(data,5)$means
test_data <- rbind(c(2,1,4,3,2,2),c(11,22,33,33,22,11), c(15,21,25,34,1,4))
k_means_predict(test_data, result)

}
