% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-medoids.R
\name{k_medoids}
\alias{k_medoids}
\title{k-medoids clustering}
\usage{
k_medoids(data, num_cluster, max_iter = 50L, tol = 1e-08)
}
\arguments{
\item{data}{data matrix. Every row contains a point}

\item{num_cluster}{int. number of clusters desired}

\item{max_iter}{(optinal) int. sets maximum number of iterations. Default: 50}

\item{tol}{(optinal) float. tolerance to conclude convergence Default: 1e-8}
}
\value{
list containing logical indicating whether the algorithm converged, number of iterations, cluster medoids
}
\description{
k-medoids clustering
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,1,0,1,2,0),ncol=2), 0.2)
k_medoids(data,4)
}
