% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{optics}
\alias{optics}
\title{optics}
\usage{
optics(data, eps, minPts)
}
\arguments{
\item{data}{The data to be clustered. A matrix, each row being a point in R^d.}

\item{eps}{The distance defining the neighborhood of a point.}

\item{minPts}{The minimum point count needed to form a dense region.}
}
\value{
Returns a list of 2 elements \code{list(ordered_list, reachability)}.\cr
The \code{ordered_list} being an indices vector, of the ordering obtained by the OPTICS algorithm and
the \code{reachability} being a vector of the reachability distances for each point of \code{data} as obtained by the OPTICS algorithm.
}
\description{
This is an implementation of the OPTICS (Ordering Points To Identify Clustering Structure) algorithm
}
\section{Details}{
 To compute the neighorhood of a data point with the \code{eps} value, this function uses the euclidean distance
}

\examples{
data <- gen_clusters(50, matrix(c(0,1,2,1,0,1,2,0),ncol=2), 0.3)
eps <- 0.5
minPts <- 3
optics(data, eps, minPts)

}
\references{
For more information visit \href{https://de.wikipedia.org/wiki/OPTICS}{this website}.
}
