% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{plot_k_means_2d}
\alias{plot_k_means_2d}
\title{Title}
\usage{
plot_k_means_2d(data, num_cluster, max_iter = 50L)
}
\arguments{
\item{max_iter}{}
}
\value{
Führt k-Means_algorithmus aus und plottet Cluster
}
\description{
Title
}
