% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{plot_k_means_spectral_clustering}
\alias{plot_k_means_spectral_clustering}
\title{plot_k_means_spectral_clustering}
\usage{
plot_k_means_spectral_clustering(data, num_cluster, dim = 2, h = 20)
}
\arguments{
\item{data}{matrix. Every row contains a point}

\item{num_cluster}{int. number of clusters desired}

\item{dim}{(optional) int. desired dimension of computed projections. Default: 2}

\item{h}{(optional) int. descent parameter of Gauss kernel}
}
\value{
plot of cluster assignment
}
\description{
visualizes results received from applying k-means algorithm on spectral projections computed by spectral-clustering
}
