% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{plot_reachability}
\alias{plot_reachability}
\title{plot_reachability}
\usage{
plot_reachability(optics_result, extract_dbscan = FALSE, eps_prime = NULL)
}
\arguments{
\item{optics_result}{result of the OPTICS algorithm (see '\code{?optics}')}

\item{extract_dbscan}{(optional) logical. runs 'extract_dbscan'. Default: FALSE}

\item{eps_prime}{(optional) value to extract clusters when running 'extract_dbscan', default is \code{eps} of optics_result.}
}
\value{
Barplot with height given by \code{reachability} of \code{optics_result} and each
bar representing a datapoint, they are orderd by \code{ordered_list} of \code{optics_result}
}
\description{
Plots the reachability-plot of the OPTICS algithm. A special kind of 'dendogram'
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,1,0,1,2,0),ncol=2), 0.3)
eps <- 0.5
minPts <- 3
#Obtain an optics result
optics_result <- optics(data, eps, minPts)
#Plot reachability:
plot_reachability(optics_result)

}
