% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{plot_spectral_clustering}
\alias{plot_spectral_clustering}
\title{plot_spectral_clustering}
\usage{
plot_spectral_clustering(data, num_cluster, dim = 2, h = 20)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthält einen Messwert in R^d}

\item{num_cluster}{int. Anzahl der Cluster}

\item{dim}{(optional) int. Dimension der erstellten k-dimensionalen Projektionen durch spektrales Clustern. Default: 2}

\item{h}{(optional) int. Abstiegsparameter des Gauss-Kerns}
}
\value{
Plot der Cluster-Zuteilung
}
\description{
veranschaulicht Ergebnis des k-means-Algortithmus angewendet auf die Daten, die durch spektrales Clustering erhalten wurden
}
