% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{plot_twoclusters_mean}
\alias{plot_twoclusters_mean}
\title{plot_twoclusters_mean}
\usage{
plot_twoclusters_mean(data, h = 20)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthält einen Messwert in R^d}

\item{h}{(optional) int. Abstiegsparameter des Gauss-Kerns}
}
\value{
plot of cluster-assignment
}
\description{
computes onedimensional projections with spectral clustering and assigns two clusters by comparing to mean.
}
