% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{spectral_clustering}
\usage{
spectral_clustering(data, dim = 2, h = 20)
}
\arguments{
\item{data}{matrix. Every row contains a point}

\item{dim}{(optional) int. desired dimension of computed projections. Default: 2}

\item{h}{(optional) int. descent parameter of Gauss kernel}
}
\value{
matrix containing k-dimensional projections (every row contains a point).
}
\description{
computes k_dimensional projections of given data by spectral-clustering
}
