% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_clusters.R
\name{gen_clusters}
\alias{gen_clusters}
\title{gen_cluster}
\usage{
gen_clusters(n, means, deviation = 0.5)
}
\arguments{
\item{n}{number of points in each cluster}

\item{means}{matrix. means of clusters
Durch die Zeilenanzahl wird die Anzahl der Cluster gesteuert.}

\item{deviation}{standard deviation. Default: 0.5}
}
\value{
returns double matrix with data
}
\description{
this function creates pseudo random data points that can be used for clustering
}
\examples{
# means <- rbind(c(0, 10), c(5, 5))
# gen_clusters(40, means)
# gen_clusters(30, means, 0.9)

}
