% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{k_means}
\alias{k_means}
\title{k-means algorithm}
\usage{
k_means(
  data,
  num_cluster,
  m0 = NULL,
  return_labels = TRUE,
  save_history = FALSE,
  max_iter = 50L,
  tol = 1e-08
)
}
\arguments{
\item{data}{matrix. Every row contains a point}

\item{num_cluster}{int. number of clusters desired}

\item{m0}{(optional) matrix with initial values. Will be chosen automatically if not specified}

\item{return_labels}{(optional) logical. returns cluster label of each instance. Default: TRUE}

\item{save_history}{(optional) logical. returns iteration history}

\item{max_iter}{(optinal) int. sets maximum number of iterations. Default: 50}

\item{tol}{(optinal) float. tolerance to conclude convergence Default: 1e-8}
}
\value{
list containing logical indicating wether the algorithm converged, number of iterations, cluster means and labels
}
\description{
Implementation of the k-means algorithm with automatic initial values (Richter, algorithm 9.10)
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,1,0,1,2,0),ncol=2), 0.3)
k_means(data,4)
data <- gen_clusters(100, matrix(c(0,0,1,1,1,0,0,1), ncol=2),0.3)
plot_k_means_2d(data,4)

}
