% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{k_means}
\alias{k_means}
\title{k-Means-Algorithmus}
\usage{
k_means(
  data,
  num_cluster,
  m0 = NULL,
  save_history = FALSE,
  return_labels = FALSE,
  max_iter = 50L,
  tol = 1e-08
)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthaelt einen Messwert in R^d}

\item{num_cluster}{int. Anzahl der Cluster}

\item{m0}{(optional) Matrix mit Anfangswerten zur Initialisierung des Algorithmus. Falls nicht angegeben werden die Startwerte automatisch gewaehlt}

\item{save_history}{(optional) logical. Gibt Cluster-Mittelpunkte und Labels in jeder Iteration zurueck. Default: FALSE}

\item{return_labels}{(optional) logical. Gibt zu jedem Messwert das Clusterlabel zurueck. Default: FALSE}

\item{max_iter}{(optinal) int. Legt maximale Anzahl an Iterationen fest. Default: 50}

\item{tol}{(optinal) float. Toleranz zur Festlegung der Konvergenz. Default: 1e-8}
}
\value{
Liste mit Konvergenznachricht, Clustermittelpunkten, sowie, falls erwuenscht Labels und Iterationen
}
\description{
Implementierung des k-Means-Algorithmus aus Richter mit zusaetzlich automatischer Startwertwahl
}
\examples{
data <- matrix(runif(100), ncol = 2); k_means(data, 5)
}
