% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{k_means_predict}
\alias{k_means_predict}
\title{k_means_predict}
\usage{
k_means_predict(x, means)
}
\arguments{
\item{x}{Ein atomarer Vektor oder eine Matrix mit Messwerten}

\item{means}{Clustermittelpunkte (z.B. aus dem k-Means-Algorithmus)}
}
\value{
erwartete Klassenlabel der neuen Messwerte
}
\description{
k_means_predict
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,0,1,2)), 0.3)
 clustering <- k_means(data,3)
 k_means_predict(c(1.2,0.8), clustering$means)

}
