% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{k_means_predict}
\alias{k_means_predict}
\title{k_means_predict}
\usage{
k_means_predict(x, means)
}
\arguments{
\item{x}{atomic vector or matrix with measurements}

\item{means}{cluster means (e.g. from k-means algorithm)}
}
\value{
predicted labels of new measurements
}
\description{
k_means_predict
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,0,1,2), ncol=2), 0.3)
 clustering <- k_means(data,3)
 k_means_predict(c(1.2,0.8), clustering$means)

}
