% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{k_means_spectral_clustering}
\alias{k_means_spectral_clustering}
\title{k_means_spectral_clustering}
\usage{
k_means_spectral_clustering(data, num_cluster, dim = 2, h = 20)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthält einen Messwert in R^d}

\item{num_cluster}{int. Anzahl der Cluster}

\item{dim}{(optional) int. Dimension der erstellten k-dimensionalen Projektionen durch spektrales Clustern. Default: 2}

\item{h}{(optional) int. Abstiegsparameter des Gauss-Kerns}
}
\value{
Liste mit Konvergenz logical, Anzahl an Iterationen, Clustermittelpunkten, sowie Labels
}
\description{
Wendet k-means-Algorithmus auf die k-dimesionalen Projektionen zu gegebenen Daten an
}
