% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-medoids.R
\name{k_medoids}
\alias{k_medoids}
\title{k_medoids}
\usage{
k_medoids(data, num_cluster, max_iter = 100L)
}
\arguments{
\item{data}{data matrix. Every row contains a point}

\item{num_cluster}{int. number of clusters desired}

\item{max_iter}{(optinal) int. sets maximum number of iterations. Default: 100}
}
\value{
list containing logical indicating whether the algorithm converged, number of iterations, cluster medoids
}
\description{
k_medoids
}
\examples{
data <- gen_clusters(50, matrix(c(0,1,2,1,0,1,2,0),ncol=2), 0.2)
k_medoids(data,4)
}
\author{
Aysegul Pekozsoy
}
