% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{optics}
\alias{optics}
\title{This is an implementation of the OPTICS Algorithm}
\usage{
optics(data, eps, minPts)
}
\arguments{
\item{data}{The data to be clustered.}

\item{eps}{The distance defining the neighborhood of a point.}

\item{minPts}{The minimum point count needed to form a dense region.}
}
\value{
Returns a list of 2 elements \code{list(ordered_list, reachability)}.
The \code{ordered_list} being an indices vector, of the ordering obtained by the OPTICS algorithm and
the \code{reachability} being a vector of the reachability distances for each point of \code{data} as obtained by the OPTICS algorithm.
}
\description{
This is an implementation of the OPTICS (Ordering Points To Identify Clustering Structure) Algorithm
}
\examples{
data <-

}
