% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means.R
\name{plot_k_means_2d}
\alias{plot_k_means_2d}
\title{plot_k_means_2d}
\usage{
plot_k_means_2d(data, num_cluster, max_iter = 50L)
}
\arguments{
\item{data}{data that should be clustered and plotted}

\item{num_cluster}{number of clusters for k-means}

\item{max_iter}{maximum iterations}
}
\value{
Performs k-means algorithm on data and plots clusters
}
\description{
plot_k_means_2d
}
