% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{plot_optics_2d}
\alias{plot_optics_2d}
\title{plot_optics_2d}
\usage{
plot_optics_2d(data, optics_result, eps_prime = optics_result$eps)
}
\arguments{
\item{data}{The data to be clustered. A matrix, each row being a point in R^d.}

\item{optics_result}{Uses the result of the OPTIC algorithm (see 'optics'), make sure it was run on \code{data}}

\item{eps_prime}{(optional) value to extract clusters, default is \code{eps} of optics_result.}
}
\value{
Plot using \code{ggplot2}
}
\description{
Plots 2-dimensional data and colors the clusters obtained by the OPTICS algorithm accordingly.
}
\section{Details}{

This function calls \code{extarct_dbscan}, hence eps_prime can be added as an optional parameter.
}

