% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{plot_spectral_clustering}
\alias{plot_spectral_clustering}
\title{plot_spectral_clustering}
\usage{
plot_spectral_clustering(data, num_cluster, f_kernel = gausskernel, dim = 2)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthält einen Messwert in R^d}

\item{num_cluster}{int. Anzahl der Cluster}

\item{f_kernel}{(optional) function. Kernfunktion. Default: Gauss-Kern mit Abstieg h}

\item{dim}{(optional) int. Dimension der erstellten k-dimensionalen Projektionen durch spektrales Clustern. Default: 2}
}
\value{
Plot der Cluster-Zuteilung
}
\description{
plot_spectral_clustering
}
