% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral-clustering.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Algorithmus: spectral_clustering}
\usage{
spectral_clustering(data, num_cluster, dim = 2, h = 50)
}
\arguments{
\item{data}{Matrix mit Daten die geclustered werden sollen. Jede Zeile enthält einen Messwert in R^d}

\item{num_cluster}{int. Anzahl der Cluster}

\item{dim}{(optional) int. Dimension der erstellten k-dimensionalen Projektionen durch spektrales Clustern. Default: 2}

\item{h}{(optional) int. Abstiegsparameter des Gauss-Kerns}
}
\value{
k-dimensionale Projektionen (selbes Format wie data)
}
\description{
berechnet k-dimensionale Projektionen zu gegebenen Daten
}
